More on finding matches between converted formats
I added code to convert the source image format to the format of the
pattern image. This is done before both images are converted to "PNG".
It helps find matches in almost all circumstances where the image type
has been changed, but there are still a few instances where it does not.