Still working to make "GIF" files work
"GIF" files cause many problems

The changes I made are:
     - if the formats of the pattern and source aren't the same, do one
of two things. If the pattern format is "GIF", convert to "RGB" and
then change format to "PNG", if not, make the source format the same as
the pattern format.
     - if the pattern format and the source format both are "GIF",
convert both to "RGB" and change both to "PNG"

These changes fix any problems with "GIF" files. There are no more
TypeErrors with them (I am pretty sure)