P mode bug
For some reason, when testing -p flower.gif -s ac1000.jpg, the source
image is in mode "P" instead of "RGB". This is causing the TypeError I
have been getting with "GIFs". I added some code so that this error
doesn't come up anymore. I am not sure yet why this is happening,
perhaps in the format changes i do in the __init__. I will look into
finding the source of the problem. The ac1000.jpg image is not in mode
"P" to begin with, so it must be changed to that  somewhere in the
program