__init__ changes
I realized that I could do what Jon did when he wrote the initial code
to change everything to "PNG"; I don't need to extract the actual name
of the file, and when I change formats, I can just save the file as
"temp_source_image.png" and save some excess computation. This way I
reduced the code significantly.

I found out that I was converting the source image to "PNG" twice under
the if statement that checks if the pattern image is a "GIF". I always
forget to write if...elif...else, which i guess makes a difference in
Python.

Anyway, the TypeError is gone and the code is cleaner. We still aren't
getting the matches we need with "GIF" files, and when analyzing
"GIFs", it takes a ridiculous amount of time to check all the pixels.
This should be the next thing we look into, because the "GIF" tests are
the only ones that don't produce matches when we need them to.