Unsupported File Formats
I added in a function to check if the files are in an acceptable
format. It does not open the files, it just checks the extension in the
string of the file location. I included 'jpg', 'jpeg', 'png', and
'gif'. If I left any out, adding these to the function will be really
easy.