tmp folder
By what i read in the emails, I guess we took out the tmp folder. I
assumed someone forgot to change the line I changed when the tmp folder
was deleted. When the program was run, this threw an error, so I made
this change and now it works for one to one images